+++ objc2-foundation/src/thread.rs
@@ -1,5 +1,3 @@
-use core::ptr::NonNull;
-
 use objc2::msg_send;
 use objc2::rc::{Id, Shared};
 use objc2::runtime::Bool;
@@ -21,8 +19,8 @@ impl NSThread {
     pub fn current() -> Id<NSThread, Shared> {
         // TODO: currentThread is @property(strong), what does that mean?
         let obj: *mut Self = unsafe { msg_send![Self::class(), currentThread] };
-        let obj = unsafe { NonNull::new_unchecked(obj) };
-        unsafe { Id::retain(obj) }
+        // TODO: Always available?
+        unsafe { Id::retain(obj).unwrap() }
     }
 
     /// Returns the [`NSThread`] object representing the main thread.
@@ -31,8 +29,7 @@ impl NSThread {
         let obj: *mut Self = unsafe { msg_send![Self::class(), mainThread] };
         // The main thread static may not have been initialized
         // This can at least fail in GNUStep!
-        let obj = NonNull::new(obj).expect("Could not retrieve main thread.");
-        unsafe { Id::retain(obj) }
+        unsafe { Id::retain(obj).expect("Could not retrieve main thread.") }
     }
 
     /// Returns `true` if the thread is the main thread.
@@ -44,7 +41,7 @@ impl NSThread {
     /// The name of the thread.
     pub fn name(&self) -> Option<Id<NSString, Shared>> {
         let obj: *mut NSString = unsafe { msg_send![self, name] };
-        unsafe { Id::retain_null(obj) }
+        unsafe { Id::retain(obj) }
     }
 }